+++ NEWS.md
@@ -275,3 +275,8 @@ This argument determines whether weights for each marker is applied when computi
 - We added argument an `n.PC` to `estPhylo` and `estNetwork` functions so that they can consider the number of PCs.
 - Some results will be changed after the update, so please install the latest version!
 - We fixed the error in `EM3.cov` function when computing random effects from the estimated variance components.
+
+
+# August 17, 2026, RAINBOWR version 0.1.42
+## Major changes (new functions, etc...)
+- We implemented new functions `calcParamSE` and `calcParamSE.cov` to calculate approximate standard errors for variance and correlation parameter estimates obtained from `EMM.cpp`, `EM3.cpp`, and `EM3.cov` using the expected Fisher information matrix.